Checkpoint
Gone to 6 bills and reduced some of the complexity of the resource files.

Unfortunately was unable to reduce the bills and the numbers of them into a single image.

So, it kinda works. Except after awhile the application terminates itself!

Needs surgery before I can return it properly, obviously.

But wanted to make sure that I had a checkpoint here.