I've now got the currency games at least usable
Kenyan currency, 6 units of currency.

Lots of fields renamed so they make sense to me

The paid box has gone into a single row. Makes things karder to see but a great deal of that is due to limited screen real estate.

Should play with margins a little to make things a little more consistent and clearer

I've greatly simplified the layout. It now sort of kind of works on different devices.

It also looks pretty awful right now (partly due to space constraints, partly due to the fact that it needs work).

It is, however, a MUCH better place to start working from. Far, far simpler and less duplication of code (starting to use styles).

At any rate, I think it is time to return